Initial release
Initial release for program TrafficSignAI that requires the German Traffic Sign Recognition Benchmark (GTSRB) data to be present in the same directory. A total of 43 signs exist in the GTSRB folder labeled 0 through 43, each containing a collection of images that have already been categorized by the road sign represented in them.